- user big decimal instead of long to store account balance and transaction amounts

+++ src/main/java/com/practical_interview/project/controllers/dtos/AccountBalanceResponse.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Builder
 @Data
 @AllArgsConstructor
@@ -15,5 +17,5 @@ public class AccountBalanceResponse {
 
     private String accountId;
 
-    private Long accountBalance;
+    private BigDecimal accountBalance;
 }
